semantic_ok

+++ src/AST/IntNode.java
@@ -4,8 +4,8 @@ import Util.IntType;
 import Util.position;
 
 public class IntNode extends constExprNode {
-    public int value;
-    IntNode(int value, position p, IntType t) {
+    public long value;
+    IntNode(long value, position p, IntType t) {
         super(p, t);
         this.value = value;
     }
